refactor: merge mage directory with catalog to keep everthing together
